python/djando-rest/src/requirements.txt
djangorestframework 
markdown 
django-filter 
coreapi